+++ test/OneImlx.Terminal.AspNetCore.Tests/TerminalHttpMapServiceTests.cs
@@ -40,63 +40,54 @@ namespace OneImlx.Terminal.AspNetCore
         public async Task RouteCommand_Processes_Command_Successfully()
         {
             // Arrange
-            var request = new TerminalRequest("test-id", "test-command");
             var context = new DefaultHttpContext();
+            using (var resposeStream = new MemoryStream())
+            {
+                context.Response.Body = resposeStream;
 
-            // Create a MemoryStream to simulate the HTTP request body with the serialized command
-            var stream = new MemoryStream();
-            await JsonSerializer.SerializeAsync(stream, request);
-            stream.Position = 0;
-            context.Request.Body = stream;
-
-            // Mock the router and processor behavior
-            mockTerminalRouter.Setup(x => x.IsRunning).Returns(true);
-            mockProcessor.Setup(x => x.IsProcessing).Returns(true);
-
-            TerminalOutput? addedResponse = null;
-            mockProcessor.Setup(x => x.ProcessRequestAsync(It.IsAny<string>(), It.IsAny<string>(), It.IsAny<string>()))
-                .Callback<string, string?, string?>((raw, senderId, senderEndpoint) =>
+                // Create a MemoryStream to simulate the HTTP request body with the serialized command
+                using (var requestStream = new MemoryStream())
                 {
-                    // Create and assign a mock response based on the input parameters
-                    addedResponse = new TerminalResponse(1, null, senderId, senderEndpoint);
-                    addedResponse.Requests[0] = new TerminalRequest("id1", raw);
-                })
-                .ReturnsAsync(() => addedResponse!);
+                    TerminalInput terminalInput = TerminalInput.Single("id1", "test-command");
+                    await JsonSerializer.SerializeAsync(requestStream, terminalInput);
+                    requestStream.Position = 0;
+                    context.Request.Body = requestStream;
+                    context.Request.ContentType = "application/json";
 
-            // Act
-            addedResponse.Should().BeNull();
-            await terminalHttpMapService.RouteBatchAsync(context);
+                    // Mock the router and processor behavior
+                    mockTerminalRouter.Setup(x => x.IsRunning).Returns(true);
+                    mockProcessor.Setup(x => x.IsProcessing).Returns(true);
 
-            // Assert
-            addedResponse.Should().NotBeNull();
-            addedResponse!.Requests.Should().HaveCount(1);
+                    TerminalOutput? addedOutput = null;
+                    mockProcessor.Setup(x => x.ExecuteAsync(It.IsAny<TerminalInput>(), It.IsAny<string>(), It.IsAny<string>()))
+                        .Callback<TerminalInput, string?, string?>((input, senderId, senderEndpoint) =>
+                        {
+                            // Create and assign a mock response based on the input parameters
+                            addedOutput = new TerminalOutput(terminalInput, ["any"], senderId, senderEndpoint);
+                        })
+                        .ReturnsAsync(() => addedOutput!);
 
-            addedResponse.Requests[0].Id.Should().Be("id1");
-            addedResponse.Requests[0].Raw.Should().Be("test-command");
-            addedResponse.BatchId.Should().BeNull();
-        }
+                    // Act
+                    addedOutput.Should().BeNull();
+                    await terminalHttpMapService.RouteAsync(context);
+                    context.Response.ContentType.Should().Be("application/json; charset=utf-8");
+                    context.Response.Body.Seek(0, SeekOrigin.Begin);
+                    using var reader = new StreamReader(context.Response.Body);
+                    string jsonResponse = await reader.ReadToEndAsync();
+                    jsonResponse.Should().Be("{\"input\":{\"batch_id\":null,\"requests\":[{\"id\":\"id1\",\"raw\":\"test-command\"}]},\"results\":[\"any\"],\"sender_endpoint\":null,\"sender_id\":null}");
 
-        // Test case to validate that a missing command string results in an exception
-        [Fact]
-        public async Task RouteCommand_Throws_When_Command_Is_Missing()
-        {
-            // Arrange
-            var request = new TerminalRequest("test-id", "  "); // Empty command string
-            var context = new DefaultHttpContext();
+                    // Assert
+                    addedOutput.Should().NotBeNull();
+                    addedOutput!.Input.Requests.Should().HaveCount(1);
 
-            // Create a MemoryStream to simulate the HTTP request body with the serialized command
-            var stream = new MemoryStream();
-            await JsonSerializer.SerializeAsync(stream, request);
-            stream.Position = 0;
-            context.Request.Body = stream;
+                    addedOutput.Input.Requests[0].Id.Should().Be("id1");
+                    addedOutput.Input.Requests[0].Raw.Should().Be("test-command");
+                    addedOutput.Input.BatchId.Should().BeNull();
 
-            // Act
-            Func<Task> act = async () => await terminalHttpMapService.RouteBatchAsync(context);
-
-            // Assert
-            await act.Should().ThrowAsync<TerminalException>()
-                .WithErrorCode("missing_command")
-                .WithErrorDescription("The command is missing in the HTTP request.");
+                    addedOutput.Results.Should().HaveCount(1);
+                    addedOutput.Results[0].Should().Be("any");
+                }
+            }
         }
 
         // Test case to validate that if the processor is not processing, the system throws an exception
@@ -104,26 +95,28 @@ namespace OneImlx.Terminal.AspNetCore
         public async Task RouteCommand_Throws_When_Processor_Is_Not_Processing()
         {
             // Arrange
-            var request = new TerminalRequest("test-id", "test-command");
+            var input = TerminalInput.Single("test-id", "test-command");
             var context = new DefaultHttpContext();
 
             // Create a MemoryStream to simulate the HTTP request body with the serialized command
-            var stream = new MemoryStream();
-            await JsonSerializer.SerializeAsync(stream, request);
-            stream.Position = 0;
-            context.Request.Body = stream;
+            using (var stream = new MemoryStream())
+            {
+                await JsonSerializer.SerializeAsync(stream, input);
+                stream.Position = 0;
+                context.Request.Body = stream;
 
-            // Mock the router to be running but the processor not processing
-            mockTerminalRouter.Setup(x => x.IsRunning).Returns(true);
-            mockProcessor.Setup(x => x.IsProcessing).Returns(false);
+                // Mock the router to be running but the processor not processing
+                mockTerminalRouter.Setup(x => x.IsRunning).Returns(true);
+                mockProcessor.Setup(x => x.IsProcessing).Returns(false);
 
-            // Act
-            Func<Task> act = async () => await terminalHttpMapService.RouteBatchAsync(context);
+                // Act
+                Func<Task> act = async () => await terminalHttpMapService.RouteAsync(context);
 
-            // Assert
-            await act.Should().ThrowAsync<TerminalException>()
-                .WithErrorCode("server_error")
-                .WithErrorDescription("The terminal processor is not processing.");
+                // Assert
+                await act.Should().ThrowAsync<TerminalException>()
+                    .WithErrorCode("server_error")
+                    .WithErrorDescription("The terminal processor is not processing.");
+            }
         }
 
         // Test case to validate that if the router is not running, the system throws an exception
@@ -131,25 +124,27 @@ namespace OneImlx.Terminal.AspNetCore
         public async Task RouteCommand_Throws_When_Router_Is_Not_Running()
         {
             // Arrange
-            var request = new TerminalRequest("test-id", "test-command");
+            var input = TerminalInput.Single("test-id", "test-command");
             var context = new DefaultHttpContext();
 
             // Create a MemoryStream to simulate the HTTP request body with the serialized command
-            var stream = new MemoryStream();
-            await JsonSerializer.SerializeAsync(stream, request);
-            stream.Position = 0;
-            context.Request.Body = stream;
+            using (var stream = new MemoryStream())
+            {
+                await JsonSerializer.SerializeAsync(stream, input);
+                stream.Position = 0;
+                context.Request.Body = stream;
 
-            // Mock the router to be not running
-            mockTerminalRouter.Setup(x => x.IsRunning).Returns(false);
+                // Mock the router to be not running
+                mockTerminalRouter.Setup(x => x.IsRunning).Returns(false);
 
-            // Act
-            Func<Task> act = async () => await terminalHttpMapService.RouteBatchAsync(context);
+                // Act
+                Func<Task> act = async () => await terminalHttpMapService.RouteAsync(context);
 
-            // Assert
-            await act.Should().ThrowAsync<TerminalException>()
-                .WithErrorCode("server_error")
-                .WithErrorDescription("The terminal HTTP router is not running.");
+                // Assert
+                await act.Should().ThrowAsync<TerminalException>()
+                    .WithErrorCode("server_error")
+                    .WithErrorDescription("The terminal HTTP router is not running.");
+            }
         }
 
         private readonly CancellationTokenSource commandTokenSource;