+++ CHANGE_LOG.md
@@ -2,8 +2,8 @@
 > Feature Change | Breaking Change
 - Introduce gRPC terminal routing
 - Introduce HTTP terminal routing
-- Introduce OneImlx.Terminal.AspNetCore for hosting terminal servers in ASP.NET Core
-- Introduce OneImlx.Terminal.Client for client application communicating with terminal servers  
+- Introduce `OneImlx.Terminal.AspNetCore` for hosting terminal servers in ASP.NET Core
+- Introduce `OneImlx.Terminal.Client` for client application communicating with terminal servers  
 
 # 5.12.3
 > Breaking Change
@@ -113,3 +113,12 @@
 # v4.3.2-rc*
 - Add TCP IP routing service
 - Port old code base and migrate to .NET Standard2.0, ,NET Standard2.1 and xUnit Tests for .NET7
+
+# v3.0.0
+- Initial release of the terminal Nuget packages
+
+# v2.0.0
+- Internal release for general framework
+
+# v1.0.0
+- Initial release for internal CLI routing